# GH-43926: [C++] Compute: RowEncoder eliminates offsets when all columns are fixed-sized