--- a/cpp/src/arrow/compute/row/row_encoder_internal.h
+++ b/cpp/src/arrow/compute/row/row_encoder_internal.h
@@ -83,6 +83,13 @@ struct ARROW_EXPORT KeyEncoder {
   static bool IsNull(const uint8_t* encoded_bytes) {
     return encoded_bytes[0] == kNullByte;
   }
+
+  struct EncoderInfo {
+    bool is_fixed_width;
+    int32_t fixed_width;
+  };
+
+  virtual EncoderInfo GetEncoderInfo() const = 0;
 };
 
 struct ARROW_EXPORT BooleanKeyEncoder : KeyEncoder {
@@ -99,6 +106,10 @@ struct ARROW_EXPORT BooleanKeyEncoder : KeyEncoder {
 
   Result<std::shared_ptr<ArrayData>> Decode(uint8_t** encoded_bytes, int32_t length,
                                             MemoryPool* pool) override;
+
+  EncoderInfo GetEncoderInfo() const override {
+    return EncoderInfo{/*is_fixed_width=*/true, 2};
+  }
 };
 
 struct ARROW_EXPORT FixedWidthKeyEncoder : KeyEncoder {
@@ -118,6 +129,10 @@ struct ARROW_EXPORT FixedWidthKeyEncoder : KeyEncoder {
   Result<std::shared_ptr<ArrayData>> Decode(uint8_t** encoded_bytes, int32_t length,
                                             MemoryPool* pool) override;
 
+  EncoderInfo GetEncoderInfo() const override {
+    return EncoderInfo{/*is_fixed_width=*/true, byte_width_ + 1};
+  }
+
   std::shared_ptr<DataType> type_;
   const int byte_width_;
 };
@@ -131,6 +146,8 @@ struct ARROW_EXPORT DictionaryKeyEncoder : FixedWidthKeyEncoder {
 
   Result<std::shared_ptr<ArrayData>> Decode(uint8_t** encoded_bytes, int32_t length,
                                             MemoryPool* pool) override;
+
+  // Uses `GetEncoderInfo` in `FixedWidthKeyEncoder`
 
   MemoryPool* pool_;
   std::shared_ptr<Array> dictionary_;
@@ -248,6 +265,10 @@ struct ARROW_EXPORT VarLengthKeyEncoder : KeyEncoder {
 
   explicit VarLengthKeyEncoder(std::shared_ptr<DataType> type) : type_(std::move(type)) {}
 
+  EncoderInfo GetEncoderInfo() const override {
+    return EncoderInfo{/*is_fixed_width=*/false, /*fixed_width=*/5};
+  }
+
   std::shared_ptr<DataType> type_;
 };
 
@@ -266,6 +287,10 @@ struct ARROW_EXPORT NullKeyEncoder : KeyEncoder {
   Result<std::shared_ptr<ArrayData>> Decode(uint8_t** encoded_bytes, int32_t length,
                                             MemoryPool* pool) override {
     return ArrayData::Make(null(), length, {NULLPTR}, length);
+  }
+
+  EncoderInfo GetEncoderInfo() const override {
+    return EncoderInfo{/*is_fixed_width=*/true, /*fixed_width=*/0};
   }
 };
 
@@ -348,18 +373,38 @@ class ARROW_EXPORT RowEncoder {
       return std::string(reinterpret_cast<const char*>(encoded_nulls_.data()),
                          encoded_nulls_.size());
     }
-    int32_t row_length = offsets_[i + 1] - offsets_[i];
-    return std::string(reinterpret_cast<const char*>(bytes_.data() + offsets_[i]),
+    int32_t row_length = 0;
+    int32_t row_offset = 0;
+    if (fixed_width_length_ != kInvalidFixedWidthOffset) {
+      row_length = fixed_width_length_;
+      row_offset = fixed_with_row_count_ * i;
+    } else {
+      row_length = offsets_[i + 1] - offsets_[i];
+      row_offset = offsets_[i];
+    }
+    return std::string(reinterpret_cast<const char*>(bytes_.data() + row_offset),
                        row_length);
   }
 
   int32_t num_rows() const {
-    return offsets_.empty() ? 0 : static_cast<int32_t>(offsets_.size() - 1);
+    if (kInvalidFixedWidthOffset == fixed_width_length_) {
+      return fixed_with_row_count_;
+    }
+    return offsets_.empty() ? 0 : offsets_[0];
   }
 
  private:
+  Status EncodeAndAppendForFixedWidth(const ExecSpan& batch);
+
+ private:
+  static constexpr int32_t kInvalidFixedWidthOffset = 1;
   ExecContext* ctx_{nullptr};
   std::vector<std::shared_ptr<KeyEncoder>> encoders_;
+  // When all columns in a row are Fixed-width or NA, the encoded row
+  // doesn't need to maintain the column offsets. In this case, the
+  // offsets_.size() would be also be empty.
+  int32_t fixed_width_length_{kInvalidFixedWidthOffset};
+  int32_t fixed_with_row_count_{0};
   // offsets_ vector stores the starting position (offset) of each encoded row
   // within the bytes_ vector. This allows for quick access to individual rows.
   //