@pitrou 2024-09-03
What does performance look like?