@mapleFU 2024-09-03
CI need fix

```
D:\a\arrow\arrow\build\cpp\src\arrow\CMakeFiles\arrow_compute_shared.dir\Unity\unity_4_cxx.cxx
D:/a/arrow/arrow/cpp/src/arrow/compute/row/row_encoder_internal.cc(359): error C2220: the following warning is treated as an error
D:/a/arrow/arrow/cpp/src/arrow/compute/row/row_encoder_internal.cc(359): warning C4244: '+=': conversion from 'const int64_t' to 'int32_t', possible loss of data
```