@mapleFU 2024-09-06
I would go out for tour from 9.14 to 9.21. I'll focus on List Join these days, mark this as wip first