@mapleFU 2024-09-08
With:
```
struct BenchmarkSettings {
  int num_threads = 1;
  JoinType join_type = JoinType::INNER;
  // Change to 'true' to benchmark alternative, non-default and less optimized version of
  // a hash join node implementation.
  bool use_basic_implementation = true;
```


After:
```
BM_HashJoinBasic_KeyTypes/"{int32}"/Hashtable krows:1        166477 ns       166453 ns         4154 rows/sec=6.15188M/s
BM_HashJoinBasic_KeyTypes/"{int32}"/Hashtable krows:8       1439597 ns      1439500 ns          487 rows/sec=5.69086M/s
BM_HashJoinBasic_KeyTypes/"{int32}"/Hashtable krows:64     13310182 ns     13310009 ns           54 rows/sec=4.92381M/s
BM_HashJoinBasic_KeyTypes/"{int32}"/Hashtable krows:512   354529054 ns    354520400 ns            2 rows/sec=1.47887M/s
BM_HashJoinBasic_KeyTypes/"{int32}"/Hashtable krows:4096 6766162518 ns   6763357800 ns            1 rows/sec=620.151k/s
BM_HashJoinBasic_KeyTypes/"{utf8}"/Hashtable krows:1         191892 ns       191763 ns         3584 rows/sec=5.33993M/s
BM_HashJoinBasic_KeyTypes/"{utf8}"/Hashtable krows:4         791714 ns       791462 ns          875 rows/sec=5.17523M/s
BM_HashJoinBasic_KeyTypes/"{utf8}"/Hashtable krows:16       3627607 ns      3627186 ns          196 rows/sec=4.517M/s
BM_HashJoinBasic_KeyTypes/"{utf8}"/Hashtable krows:64      22743811 ns     22741736 ns           33 rows/sec=2.88175M/s
```

Before:
```
BM_HashJoinBasic_KeyTypes/"{int32}"/Hashtable krows:1        162843 ns       162781 ns         4207 rows/sec=6.29067M/s
BM_HashJoinBasic_KeyTypes/"{int32}"/Hashtable krows:8       1421683 ns      1421186 ns          493 rows/sec=5.7642M/s
BM_HashJoinBasic_KeyTypes/"{int32}"/Hashtable krows:64     13439032 ns     13438304 ns           52 rows/sec=4.87681M/s
BM_HashJoinBasic_KeyTypes/"{int32}"/Hashtable krows:512   408234871 ns    408206400 ns            2 rows/sec=1.28437M/s
BM_HashJoinBasic_KeyTypes/"{int32}"/Hashtable krows:4096 6854617796 ns   6852387200 ns            1 rows/sec=612.094k/s
BM_HashJoinBasic_KeyTypes/"{utf8}"/Hashtable krows:1         188940 ns       188805 ns         3716 rows/sec=5.42359M/s
BM_HashJoinBasic_KeyTypes/"{utf8}"/Hashtable krows:4         759144 ns       758925 ns          918 rows/sec=5.39711M/s
BM_HashJoinBasic_KeyTypes/"{utf8}"/Hashtable krows:16       3477411 ns      3476909 ns          202 rows/sec=4.71223M/s
BM_HashJoinBasic_KeyTypes/"{utf8}"/Hashtable krows:64      20586176 ns     20585156 ns           34 rows/sec=3.18365M/s
```